+++ system/pr_status_syscall_summary.c
@@ -32,7 +32,7 @@ void pr_status_syscall_summary(){
     for(i=0;i<=max_pid;i++){        
         for(j=create_enum; j< number_of_system_calls; j++){
             if(procsumm_table[i].rec_count[j]!=0){
-                int average_cycles= procsumm_table[i].total_cycles[j]/procsumm_table[i].rec_count[j];
+                double average_cycles= procsumm_table[i].total_cycles[j]/procsumm_table[i].rec_count[j];
                 printf("%3d %7s %5d %-14d\n",
                         i, sys_call_description[j],procsumm_table[i].rec_count[j],average_cycles);
